+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Releases
 
+### v0.0.4 (2025-09-24)
+
 - **Added:** Sort warning by line number.
 - **Added:** Option `--statistics`: to show counts for every rule with at least one violation.
 - **Fixed:** Bug in walking directories when using `--ignore` option.